:sparkles: board CRUD security 추가

+++ src/main/java/com/rooi/rooi/service/BoardService.java
@@ -4,11 +4,12 @@ import com.rooi.rooi.dto.ApiResponseDto;
 import com.rooi.rooi.dto.BoardRequestDto;
 import com.rooi.rooi.dto.BoardResponseDto;
 import com.rooi.rooi.entity.Board;
+import com.rooi.rooi.entity.User;
+import com.rooi.rooi.repository.BoardRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-import com.rooi.rooi.repository.BoardRepository;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -29,30 +30,36 @@ public class BoardService {
 		return new BoardResponseDto(board);
 	}
 
-	public BoardResponseDto createBoard(BoardRequestDto requestDto) {
+	public BoardResponseDto createBoard(BoardRequestDto requestDto, User user) {
 		log.info("boardService - createBoard");
-		Board board = boardRepository.save(new Board(requestDto));
+		Board board = boardRepository.save(new Board(requestDto, user));
 		return new BoardResponseDto(board);
 	}
 
 	@Transactional
-	public BoardResponseDto updateBoard(Long id, BoardRequestDto requestDto) {
+	public BoardResponseDto updateBoard(Long id, BoardRequestDto requestDto, User user) {
+		// 존재 여부 확인
 		Board board = findBoard(id);
 
-		// TODO : user 추가 후 if문으로 작성자 확인 로직 추가
-		board.setTitle(requestDto.getTitle());
-		board.setContents(requestDto.getContests());
-		board.setBoardColor(requestDto.getBoardColor());
-
+		if (board.getUser().getId().equals(user.getId())) {
+			board.setTitle(requestDto.getTitle());
+			board.setContents(requestDto.getContests());
+			board.setBoardColor(requestDto.getBoardColor());
+		} else {
+			throw new IllegalArgumentException("작성자만 수정할 수 있습니다.");
+		}
 		return new BoardResponseDto(board);
 	}
 
-	public ApiResponseDto deleteBoard(Long id) {
+	public ApiResponseDto deleteBoard(Long id, User user) {
+		// 존재 여부 확인
 		Board board = findBoard(id);
 
-		// TODO : user 추가 후 if문으로 작성자 확인 로직 추가
-		boardRepository.delete(board);
-
+		if (board.getUser().getId().equals(user.getId())) {
+			boardRepository.delete(board);
+		} else {
+			throw new IllegalArgumentException("작성자만 삭제할 수 있습니다.");
+		}
 		return new ApiResponseDto("보드 삭제 완료", HttpStatus.OK.value());
 	}
